At least it compiles

+++ headers/line.h
@@ -10,18 +10,18 @@ class Line {
 private:
     string text;
     // the delay, in ms, from the start of the song
-    uint_fast64_t delay_ms;
+    size_t delay_ms;
 
 public:
     Line();
     Line(string ln);
-    Line(string ln, uint_fast64_t delay);
+    Line(string ln, size_t delay);
     // setters
     bool set_string(string s);
-    bool set_delay(uint_fast64_t delay);
+    bool set_delay(size_t delay);
     // getters
     string get_text();
-    uint_fast64_t get_delay();
+    size_t get_delay();
 };
 
 #endif
